Refactor and optimize codebase for consistency
- Improved parameter passing in `Actor.h` to avoid unnecessary copies.
- Fixed constructor initialization order in `Enemy.h` and `Player.h`.
- Refactored `Engine.h` singleton pattern and increased particle pool size.
- Added `RemoveActors` method and pending actor handling in `Scene`.
- Updated `ParticleSystem` and `Renderer` for type safety and clarity.
- Replaced hardcoded assets in `Assets.cpp` with C++17 inline variables.
- Standardized math constants and functions in `mathUitl.h`.
- Added object lifecycle management demo in `Game.cpp`.
- Updated project files for better build configuration support.
- Fixed unresolved external symbol for `Player::Draw`.
- Replaced `sqrt` with `std::sqrt` in `Vector2.h` and `Vector3.h`.
- Updated `Text` dimensions to use `float` for better precision.

+++ Engine/ParticleSystem.cpp
@@ -29,7 +29,7 @@ namespace nu
 
 			
 			float angle = desc.angle + RandomFloat(-desc.angleVariance, desc.angleVariance);
-			float speed = desc.speed + RandomFloat(-desc.speedVariance, desc.speedVariance);
+			float speed = static_cast<float>(desc.speed) + RandomFloat(-desc.speedVariance, desc.speedVariance);
 
 			particles->position = desc.position;
 			particles->velocity = Vector2{ std::cos(angle), std::sin(angle) } * speed;
